### 2022-11-25 #4964 @editorialbot
[PRE REVIEW]: HiPart: Hierarchical divisive clustering toolbox
**Submitting author:** @panagiotisanagnostou (<a href="http://orcid.org/0000-0002-4775-9220">Panagiotis Anagnostou</a>)
**Repository:** https://github.com/panagiotisanagnostou/HiPart
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.18
