### 2022-12-08 #4964 @jbytecode
👋👋👋 Dear @jonjoncardoso   👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**HiPart: Hierarchical divisive clustering toolbox**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4964).

Thank you in advance!
