### 2022-12-14 #4964 @gagolews
Apologies, I am too busy.

My I suggest, though, that the authors (@panagiotisanagnostou ...) include, at some point, the ITM algorithm <https://github.com/amueller/information-theoretic-mst> (see also https://link.springer.com/chapter/10.1007/978-3-642-32717-9_21) - a divisive scheme based on minimum spanning trees; tends to work pretty well on a range of benchmark datasets (see https://clustering-benchmarks.gagolewski.com/ ...)
