### 2022-12-21 #4964 @jbytecode
👋👋👋 Dear @jjerphan   👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**HiPart: Hierarchical divisive clustering toolbox**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4964).

We have already assigned the first reviewer, we are seeking for the second one!

Thank you in advance!
