### 2022-12-21 #4964 @adavidzh
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**HiPart: Hierarchical divisive clustering toolbox**". You can find more information at the top of this Github issue (#4964).

Hey @jbytecode, I'm afraid that I cannot take this on at this time or for the coming month.
