### 2022-12-29 fa7c87b
MAINT Remove requirements.txt file
This comes as a duplicate with respect to the list given
in setup.py.
