+++ requirements.txt
@@ -1,10 +1,0 @@
-numpy
-treelib >= 1.6
-scipy
-scikit-learn
-statsmodels>=0.13
-kdepy
-matplotlib
-plotly
-dash >= 2.0
-pandas
