### 2022-12-30 #5024 @jjerphan
I started reviewing the HiPart in the last few days.

I am drafting https://github.com/panagiotisanagnostou/HiPart/pull/19 to provide explicit suggestions for the repository.
