### 2023-01-03 c30da19
Update python versions support.
We will stop the continuous integration checks for this python version because of the imminent stop of support for python version 3.7, according to PEP 537. The support for python version 3.7 will stop on the 6th of 2023.
