+++ src/HiPart/__init__.py
@@ -37,7 +37,7 @@ from KDEpy.NaiveKDE import NaiveKDE
 from KDEpy.TreeKDE import TreeKDE
 from KDEpy.FFTKDE import FFTKDE
 
-__version__ = "0.1.18"
+__version__ = "0.2.0"
 __author__ = "Panagiotis Anagnostou"
 
 TreeKDE = TreeKDE
