+++ tests/test_package.py
@@ -1,6 +1,5 @@
 from __future__ import unicode_literals
 from distutils import dir_util
-from HiPart import __utility_functions as util
 from HiPart import visualizations as viz
 from HiPart.clustering import dePDDP
 from HiPart.clustering import iPDDP
