+++ docs/conf.py
@@ -21,7 +21,7 @@ copyright = '2022, Panagiotis Anagnostou'
 author = 'Panagiotis Anagnostou'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.17'
+release = '0.2.0'
 
 # -- General configuration ---------------------------------------------------
 
