+++ README.md
@@ -11,7 +11,7 @@ This repository presents the HiPart package, an open-source native python librar
 
 Installation
 ------------
-For the installation of the package, the only necessary actions and requirements are a version of Python higher or equal to 3.7 and the execution of the following command.
+For the installation of the package, the only necessary actions and requirements are a version of Python higher or equal to 3.8 and the execution of the following command.
 
 ```bash
 pip install HiPart
