+++ setup.py
@@ -21,15 +21,16 @@ setuptools.setup(
     include_package_data=True,
     python_requires=">=3.7",
     install_requires=[
-        "numpy",
         "treelib>=1.6",
-        "scipy",
-        "scikit-learn",
+        "scipy==1.7",
         "statsmodels>=0.13",
         "kdepy",
-        "matplotlib",
+        "matplotlib==3.5",
         "plotly",
         "dash>=2.0",
+        "numpy==1.21",
+        "scikit-learn==1.0",
+        "pandas==1.3",
     ],
     project_urls={
         "Bug Tracker": "https://github.com/panagiotisanagnostou/HiPart",
