+++ docs/index.rst
@@ -25,22 +25,22 @@ Installation
 ------------
 For the installation of the package, the only necessary actions and requirements are a version of Python higher or equal to 3.7 and the execution of the following command.
 
-	.. code-block:: sh
-	
-		pip install HiPart
+   .. code-block:: sh
+   
+      pip install HiPart
 
 Simple Example Execution
 ------------------------
 The example bellow is the simplest form of the package's execution. Shortly, it shows the creation of synthetic clustering dataset containing 6 clusters. Afterwards it is clustered with the dePDDP algorithm and only the cluster labels are returned.
 
-	.. code-block:: py
-	
-		from HiPart.clustering import dePDDP
-		from sklearn.datasets import make_blobs
+   .. code-block:: py
+   
+      from HiPart.clustering import dePDDP
+      from sklearn.datasets import make_blobs
 
-		X, y = make_blobs(n_samples=1500, centers=6, random_state=0)
+      X, y = make_blobs(n_samples=1500, centers=6, random_state=0)
 
-		clustered_class = dePDDP(max_clusters_number=6).fit_predict(X)
+      clustered_class = dePDDP(max_clusters_number=6).fit_predict(X)
 
 Users can find complete execution examples for all the algorithms of the HiPart package in the `clustering_example <https://github.com/panagiotisanagnostou/HiPart/blob/main/examples/clustering_example.py>`_ file of the repository. Also, the users can find a KernelPCA method usage example in the `clustering_with_kpca_example <https://github.com/panagiotisanagnostou/HiPart/blob/main/examples/clustering_with_kpca_example.py>`_ file of the repository. Finally, the file `interactive_visualization_example <https://github.com/panagiotisanagnostou/HiPart/blob/main/examples/interactive_visualization_example.py>`_ contains an example execution of the interactive visualization. The instructions for the interactive visualization GUI can be found with the execution of this visualization.
 
