### 2023-01-14 38d17c6
dePDDP execution time enhancement
Change the way the dePPDP algorithm calculates the local minimum points on the density of the projected data. Minor update on the pytests to take into account this change.
