### 2023-01-14 87e49f3
Clarify remarks about class instances
