### 2023-01-14 #5024 @jjerphan
### A remark of Conflict of interest

I am one of the maintainers of scikit-learn which is providing reference implementation of clustering algorithms. I do not have any conflict of interest in the favor of the acceptance of the rejection of this paper: scikit-learn contributors and maintainers welcome qualitative third party packages extending scikit-learn in the ecosystem.

### Last review items' clarification

As of https://github.com/panagiotisanagnostou/HiPart/tree/f6c0d3785741f752a5a2cecafcef1d1eb03347bb, I have not checked the following 7 items of the 21 items mainly because I have questions, remarks or suggestions to improve of them:

 - Contribution and authorship
 - Substantial scholarly effort
 - Functionality
 - Performance
 - Reproducibility
 - Community guidelines
 - Quality of writing

Bellow are my questions, remarks or suggestions for each of them.

---

> - **Contribution and authorship**: Has the submitting author (@panagiotisanagnostou) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

It is clear that @panagiotisanagnostou is the main major contributor to HiPart. @stevestavropoulos also has contributed notable to HiPart.

To validate this item I would like to know the rationale for other authors' presence given that authors are indicated as having contributed equally.

> - **Substantial scholarly effort**: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

I think HiPart is rather small as a package with perfectible aspects. To me, HiPart currently do not meet all the criteria for being eligible to be published in JOSS. Yet, I think it is on track and that this submission is appreciable and legitimate.

Perfectible aspects of HiPart are:
 - its documentation: for now, there is a small number of undocumented examples. Notebook styles examples for all methods might improve HiPart approachability and adoption.
 - names containing typos and not PEP8-compliant: as an example, I think `bicecting_kmeans` must be renamed `BisectingKmeans`
 - lack comparisons with other libraries in the Scientific Python Ecosystem, in particular proof of functionality for the targeted problems (scalable methods for high dimensional data clustering) with respect to alternatives.
 - code quality: there's currently a lot of duplication on some files (see the ones graded as F). Due to the size of the library, some more effort can bring it to a A grade on Codacy.

> **Functionality**: Have the functional claims of the software been confirmed?
> **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
> - **Reproducibility**: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

I unfortunately can't reproduce results (benchmarks and figures) using script in [`paper/paper_scripts`](https://github.com/panagiotisanagnostou/HiPart/tree/paper/paper_scripts). Some import paths are invalid and data is missing.

To validate this item, please do check that those scripts work (I think using relative paths might help prevent issues).

Moreover, from my reading of the implementations (which are mainly backed by NumPy), I hardly think implementations can be scalable or efficient for the targeted problems.

I can't say if HiPart is completely functional, is performant, and if claims are backed by reproducible results.

> - **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Guidelines for 2) and 3) are clear to me but guidelines for 1) are missing. Adding a simple contributing guide as `CONTRIBUTING.md` would suffice to me.

> - **Quality of writing**: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

See https://github.com/panagiotisanagnostou/HiPart/pull/23 for remarks for this items.
