### 2023-01-20 e2b6f89
paper_scripts and documentation update
