+++ docs/examples.rst
@@ -1,0 +1,7 @@
+Advanced Examples
+=================
+
+.. toctree::
+   :maxdepth: 3
+
+   hipart-example1
