### 2023-01-20 f87d5b1
Update HiPart.rst
