+++ docs/examples.rst
@@ -1,7 +1,0 @@
-Advanced Examples
-=================
-
-.. toctree::
-   :maxdepth: 3
-
-   hipart-example1
