+++ examples/clustering_with_kpca_example.py
@@ -2,7 +2,7 @@ import HiPart.visualizations as viz
 import matplotlib.pyplot as plt
 import numpy as np
 
-from HiPart.clustering import dePDDP
+from HiPart.clustering import DePDDP
 from sklearn.decomposition import KernelPCA
 from sklearn.datasets import make_circles
 
@@ -80,7 +80,7 @@ if __name__ == "__main__":
 
     clusternumber = 2
 
-    outObj = dePDDP(
+    outObj = DePDDP(
         decomposition_method="kpca",
         max_clusters_number=clusternumber,
         bandwidth_scale=0.5,
