+++ tests/test_package.py
@@ -1,11 +1,11 @@
 from __future__ import unicode_literals
 from distutils import dir_util
 from HiPart import visualizations as viz
-from HiPart.clustering import dePDDP
-from HiPart.clustering import iPDDP
-from HiPart.clustering import kM_PDDP
+from HiPart.clustering import DePDDP
+from HiPart.clustering import IPDDP
+from HiPart.clustering import KMPDDP
 from HiPart.clustering import PDDP
-from HiPart.clustering import bicecting_kmeans
+from HiPart.clustering import BisectingKmeans
 
 import numpy as np
 import os
@@ -15,11 +15,11 @@ import pytest
 
 @pytest.fixture
 def datadir(tmpdir, request):
-    '''
+    """
     Fixture responsible for searching a folder with the same name of test
     module and, if available, moving all contents to a temporary directory so
     tests can use them freely.
-    '''
+    """
     filename = request.module.__file__
     test_dir, _ = os.path.splitext(filename)
 
@@ -29,28 +29,28 @@ def datadir(tmpdir, request):
     return tmpdir
 
 
-def test_dePDDP_return_type(datadir):
+def test_DePDDP_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
-    new_obj = dePDDP(max_clusters_number=5).fit(data_import["data"])
-    assert isinstance(new_obj, dePDDP)
+    new_obj = DePDDP(max_clusters_number=5).fit(data_import["data"])
+    assert isinstance(new_obj, DePDDP)
 
 
-def test_iPDDP_return_type(datadir):
+def test_IPDDP_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
-    new_obj = iPDDP(max_clusters_number=5).fit(data_import["data"])
-    assert isinstance(new_obj, iPDDP)
+    new_obj = IPDDP(max_clusters_number=5).fit(data_import["data"])
+    assert isinstance(new_obj, IPDDP)
 
 
-def test_kM_PDDP_return_type(datadir):
+def test_KMPDDP_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
-    new_obj = kM_PDDP(max_clusters_number=5).fit(data_import["data"])
-    assert isinstance(new_obj, kM_PDDP)
+    new_obj = KMPDDP(max_clusters_number=5).fit(data_import["data"])
+    assert isinstance(new_obj, KMPDDP)
 
 
 def test_PDDP_return_type(datadir):
@@ -65,49 +65,49 @@ def test_bicecting_kmeans_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
-    new_obj = bicecting_kmeans(max_clusters_number=5).fit(data_import["data"])
-    assert isinstance(new_obj, bicecting_kmeans)
+    new_obj = BisectingKmeans(max_clusters_number=5).fit(data_import["data"])
+    assert isinstance(new_obj, BisectingKmeans)
 
 
-def test_dePDDP_parameter_errors():
+def test_DePDDP_parameter_errors():
     success_score = 0
 
-    algorithm = dePDDP()
+    algorithm = DePDDP()
     success_score += 1 if isinstance(algorithm.decomposition_method, str) else 0
     success_score += 1 if isinstance(algorithm.max_clusters_number, int) else 0
-    success_score += 1 if isinstance(algorithm.split_data_bandwidth_scale, float) else 0
+    success_score += 1 if isinstance(algorithm.bandwidth_scale, float) else 0
[…]

[…]
