### 2023-01-30 ec317c7
Delete tester.py
