### 2023-01-31 1f92bec
Update on HiPart API, source , and documentation
In this commit we updated:

* The naming of our classes
* The effectiveness of the design for our algorithms
* We fixed a tuple of typos and added necessary field in the documentation of the package

Co-Authored-By: usersotiris <38103925+usersotiris@users.noreply.github.com>
