+++ setup.py
@@ -3,7 +3,7 @@ import setuptools
 with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 setuptools.setup(
     name="HiPart",
