+++ src/HiPart/clustering.py
@@ -68,8 +68,13 @@ class DePDDP:
     ----------
     output_matrix : numpy.ndarray
         Model's step by step execution output.
-    labels_ :
-        Extracted clusters from the algorithm
+    labels_ : numpy.ndarray
+        Extracted clusters from the algorithm.
+    tree : treelib.Tree
+        The object which contains all the information about the execution of
+        the dePDDP algorithm.
+    samples_number : int
+        The number of samples contained in the data.
 
     """
 
@@ -81,7 +86,7 @@ class DePDDP:
         percentile=0.1,
         min_sample_split=5,
         visualization_utility=True,
-        **decomposition_args
+        **decomposition_args,
     ):
         self.decomposition_method = decomposition_method
         self.max_clusters_number = max_clusters_number
@@ -110,43 +115,43 @@ class DePDDP:
 
         """
         self.X = X
-        self.samples_number = X.shape[0]
+        self.samples_number = np.size(X, 0)
 
         # create an id vector for the samples of X
-        indices = np.array([int(i) for i in range(np.size(self.X, 0))])
+        indices = np.array([int(i) for i in range(self.samples_number)])
 
-        # initialize tree and root node                 # step (0)
-        tree = Tree()
+        # initialize the tree and root node                           # step (0)
+        den_tree = Tree()
         # nodes unique IDs indicator
         self.node_ids = 0
         # nodes next color indicator (necessary for visualization purposes)
         self.cluster_color = 0
-        tree.create_node(
-            tag="cl_" + str(self.node_ids),
+        den_tree.create_node(
+            tag="density_cluster_" + str(self.node_ids),
             identifier=self.node_ids,
-            data=self.calculate_node_data(indices, self.X, self.cluster_color),
+            data=self.calculate_node_data(indices, self.cluster_color),
         )
         # indicator for the next node to split
         selected_node = 0
 
-        # if no possibility of split exists on the data     # (ST2)
-        if not tree.get_node(0).data["split_permission"]:
-            raise RuntimeError("cannot split the data at all!!!")
+        # if no possibility of split exists on the data                  # (ST2)
+        if not den_tree.get_node(0).data["split_permission"]:
+            raise RuntimeError("dePDDP cannot split the data at all!!!")
 
         # Initialize the ST1 stopping criterion counter that count the number
-        # of clusters                                       # (ST1)
+        # of clusters                                                    # (ST1)
         found_clusters = 1
-        while (selected_node is not None) and (
-            found_clusters < self.max_clusters_number
+        while (found_clusters < self.max_clusters_number) and (
+            selected_node is not None
         ):  # (ST1) or (ST2)
 
-            self.split_function(tree, selected_node)  # step (1)
+            self.split_function(den_tree, selected_node)  # step (1, 2)
 
             # select the next kid for split based on the local minimum density
-            selected_node = self.select_kid(tree.leaves())  # step (2)
+            selected_node = self.select_kid(den_tree.leaves())  # step (3)
             found_clusters = found_clusters + 1  # (ST1)
 
-        self.tree = tree
+        self.tree = den_tree
         return self
 
     def fit_predict(self, X):
@@ -169,7 +174,7 @@ class DePDDP:
 
         return self.fit(X).labels_
 
-    def split_function(self, tree, selected_node):
+    def split_function(self, den_tree, selected_node):
         """
         Split the indicated node on the minimum of the local minimum density
         of the data projected on the first principal component.
@@ -179,11 +184,11 @@ class DePDDP:
 
         Parameters
         ----------
-        selected_node : int
-            The numerical identifier for the tree node that i about to be split.
-        tree : treelib.tree.Tree
[…]
