+++ tests/test_package.py
@@ -29,7 +29,7 @@ def datadir(tmpdir, request):
     return tmpdir
 
 
-def test_DePDDP_return_type(datadir):
+def test_depddp_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -37,7 +37,7 @@ def test_DePDDP_return_type(datadir):
     assert isinstance(new_obj, DePDDP)
 
 
-def test_IPDDP_return_type(datadir):
+def test_ipddp_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -45,7 +45,7 @@ def test_IPDDP_return_type(datadir):
     assert isinstance(new_obj, IPDDP)
 
 
-def test_KMPDDP_return_type(datadir):
+def test_kmpddp_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -53,7 +53,7 @@ def test_KMPDDP_return_type(datadir):
     assert isinstance(new_obj, KMPDDP)
 
 
-def test_PDDP_return_type(datadir):
+def test_pddp_return_type(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -69,7 +69,7 @@ def test_bicecting_kmeans_return_type(datadir):
     assert isinstance(new_obj, BisectingKmeans)
 
 
-def test_DePDDP_parameter_errors():
+def test_depddp_parameter_errors():
     success_score = 0
 
     algorithm = DePDDP()
@@ -104,7 +104,7 @@ def test_DePDDP_parameter_errors():
     assert success_score == 11
 
 
-def test_IPDDP_parameter_errors(datadir):
+def test_ipddp_parameter_errors(datadir):
     success_score = 0
 
     algorithm = IPDDP()
@@ -134,7 +134,7 @@ def test_IPDDP_parameter_errors(datadir):
     assert success_score == 9
 
 
-def test_KMPDDP_parameter_errors(datadir):
+def test_kmpddp_parameter_errors(datadir):
     success_score = 0
 
     algorithm = KMPDDP(random_seed=123)
@@ -164,7 +164,7 @@ def test_KMPDDP_parameter_errors(datadir):
     assert success_score == 9
 
 
-def test_PDDP_parameter_errors(datadir):
+def test_pddp_parameter_errors(datadir):
     success_score = 0
 
     algorithm = PDDP()
@@ -213,7 +213,7 @@ def test_bicecting_kmeans_parameter_errors():
     assert success_score == 6
 
 
-def test_DePDDP_labels__return_type_and_form(datadir):
+def test_depddp_labels__return_type_and_form(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -221,7 +221,7 @@ def test_DePDDP_labels__return_type_and_form(datadir):
     assert isinstance(results, np.ndarray) and results.ndim == 1
 
 
-def test_IPDDP_labels__return_type_and_form(datadir):
+def test_ipddp_labels__return_type_and_form(datadir):
     with open(datadir.join('test_data.dump'), "rb") as inf:
         data_import = pickle.load(inf)
 
@@ -229,7 +229,7 @@ def test_IPDDP_labels__return_type_and_form(datadir):
     assert isinstance(results, np.ndarray) and results.ndim == 1
 
 
-def test_KMPDDP_labels__return_type_and_form(datadir):
+def test_kmpddp_labels__return_type_and_form(datadir):
[…]
