+++ src/HiPart/clustering.py
@@ -475,7 +475,7 @@ class DePDDP:
     @output_matrix.setter
     def output_matrix(self, v):
         raise RuntimeError(
-            "DePDDP: output_matrix: can only be generated and not to be created!"
+            "DePDDP: output_matrix: can only be generated and not to be assigned!"
         )
 
     @property
@@ -483,13 +483,12 @@ class DePDDP:
         labels_ = np.ones(np.size(self.X, 0))
         for i in self.tree.leaves():
             labels_[i.data["indices"]] = i.identifier
-        labels_
         return labels_
 
     @labels_.setter
     def labels_(self, v):
         raise RuntimeError(
-            "DePDDP: labels_: can only be generated and not to be created!"
+            "DePDDP: labels_: can only be generated and not to be assigned!"
         )
 
 
@@ -914,7 +913,7 @@ class IPDDP:
     @output_matrix.setter
     def output_matrix(self, v):
         raise RuntimeError(
-            "IPDDP: output_matrix: can only be generated and not to be created!"
+            "IPDDP: output_matrix: can only be generated and not to be assigned!"
         )
 
     @property
@@ -927,7 +926,7 @@ class IPDDP:
     @labels_.setter
     def labels_(self, v):
         raise RuntimeError(
-            "IPDDP: labels_: can only be generated and not to be created!"
+            "IPDDP: labels_: can only be generated and not to be assigned!"
         )
 
 
@@ -1331,13 +1330,11 @@ class KMPDDP:
                 # create output cluster splitting matrix
                 tmp = np.copy(output_matrix[-1])
                 # Left child according to the tree creation process
-                tmp[self.tree.children(j)[0].data["indices"]] = self.tree.children(j)[
-                    0
-                ].identifier
+                tmp[self.tree.children(j)[0].data["indices"]] =\
+                    self.tree.children(j)[0].identifier
                 # Right child according to the tree creation process
-                tmp[self.tree.children(j)[1].data["indices"]] = self.tree.children(j)[
-                    1
-                ].identifier
+                tmp[self.tree.children(j)[1].data["indices"]] =\
+                    self.tree.children(j)[1].identifier
 
                 # The output_matrix is created transposed
                 output_matrix.append(tmp)
@@ -1352,7 +1349,7 @@ class KMPDDP:
     @output_matrix.setter
     def output_matrix(self, v):
         raise RuntimeError(
-            "KMPDDP: output_matrix: can only be generated and not to be created!"
+            "KMPDDP: output_matrix: can only be generated and not to be assigned!"
         )
 
     @property
@@ -1365,7 +1362,7 @@ class KMPDDP:
     @labels_.setter
     def labels_(self, v):
         raise RuntimeError(
-            "KMPDDP: labels_: can only be generated and not to be created!"
+            "KMPDDP: labels_: can only be generated and not to be assigned!"
         )
 
 
@@ -1458,10 +1455,10 @@ class PDDP:
         tree.create_node(
             tag="cl_" + str(self.node_ids),
             identifier=self.node_ids,
-            data=self.calculate_node_data(indices, self.X, self.cluster_color),
+            data=self.calculate_node_data(indices, self.cluster_color),
         )
         # indicator for the next node to split
-        selected_node = 0
+        selected = 0
 
         # if no possibility of split exists on the data     # (ST2)
         if not tree.get_node(0).data["split_permission"]:
@@ -1469,16 +1466,16 @@ class PDDP:
 
         # Initialize the ST1 stopping criterion counter that count the number
         # of clusters                                       # (ST1)
-        found_clusters = 1
-        while (selected_node is not None) and (
-            found_clusters < self.max_clusters_number
+        counter = 1
+        while (selected is not None) and (
+            counter < self.max_clusters_number
         ):  # (ST1) or (ST2)
 
[…]
