+++ CODE_OF_CONDUCT.md
@@ -17,28 +17,28 @@ diverse, inclusive, and healthy community.
 Examples of behavior that contributes to a positive environment for our
 community include:
 
-*  Demonstrating empathy and kindness toward other people
+-   Demonstrating empathy and kindness toward other people
 
-*  Being respectful of differing opinions, viewpoints, and experiences
+-   Being respectful of differing opinions, viewpoints, and experiences
 
-*  Giving and gracefully accepting constructive feedback
+-   Giving and gracefully accepting constructive feedback
 
-*  Accepting responsibility and apologizing to those affected by our mistakes, and learning from the experience
+-   Accepting responsibility and apologizing to those affected by our mistakes, and learning from the experience
 
-*  Focusing on what is best not just for us as individuals, but for the
-overall community
+-   Focusing on what is best not just for us as individuals, but for the
+    overall community
 
 Examples of unacceptable behavior include:
 
-*  The use of sexualized language or imagery, and sexual attention or advances of any kind
+-   The use of sexualized language or imagery, and sexual attention or advances of any kind
 
-*  Trolling, insulting or derogatory comments, and personal or political attacks
+-   Trolling, insulting or derogatory comments, and personal or political attacks
 
-*  Public or private harassment
+-   Public or private harassment
 
-*  Publishing others' private information, such as a physical or email address, without their explicit permission
+-   Publishing others' private information, such as a physical or email address, without their explicit permission
 
-*  Other conduct which could reasonably be considered inappropriate in a professional setting
+-   Other conduct which could reasonably be considered inappropriate in a professional setting
 
 ## Enforcement Responsibilities
 
@@ -119,13 +119,11 @@ the community.
 
 This Code of Conduct is adapted from the Contributor Covenant: homepage,
 version 2.0, available at
-https://www.contributor-covenant.org/version/2/0/code_of_conduct.html.
+.
 
 Community Impact Guidelines were inspired by [Mozilla's code of conduct
 enforcement ladder](https://github.com/mozilla/diversity).
 
 [homepage](https://www.contributor-covenant.org)
 
-For answers to common questions about this code of conduct, see the FAQ at
-https://www.contributor-covenant.org/faq. Translations are available at
-https://www.contributor-covenant.org/translations.
+For answers to common questions about this code of conduct, see the FAQ at.

[…]
