### 2023-01-31 e3b2368
Typos and duplication reduction
