+++ CODE_OF_CONDUCT.md
@@ -31,13 +31,9 @@ community include:
 Examples of unacceptable behavior include:
 
 -   The use of sexualized language or imagery, and sexual attention or advances of any kind
-
 -   Trolling, insulting or derogatory comments, and personal or political attacks
-
 -   Public or private harassment
-
 -   Publishing others' private information, such as a physical or email address, without their explicit permission
-
 -   Other conduct which could reasonably be considered inappropriate in a professional setting
 
 ## Enforcement Responsibilities
