+++ src/HiPart/clustering.py
@@ -228,7 +228,7 @@ class DePDDP:
         self.cluster_color += 1
         self.node_ids += 2
 
-    def select_kid(self, leaves):
+    def select_kid(self, possible_splits):
         """
         The clusters each time exist in the leaves of the trees. From those
         leaves select the next leave to split based on the algorithm's
@@ -239,7 +239,7 @@ class DePDDP:
 
         Parameters
         ----------
-        leaves : list of treelib.node.Node
+        possible_splits : list of treelib.node.Node
             The list of nodes needed to exam to select the next Node to split.
 
         Returns
@@ -251,18 +251,18 @@ class DePDDP:
         min_density_node = None
 
         # Remove the nodes that can not split further
-        leaves = list(
-            np.array(leaves)[
+        possible_splits = list(
+            np.array(possible_splits)[
                 [
                     True if i.data["split_criterion"] is not None else False
-                    for i in leaves
+                    for i in possible_splits
                 ]
             ]
         )
 
-        if len(leaves) > 0:
+        if len(possible_splits) > 0:
             for i in sorted(
-                enumerate(leaves), key=lambda x: x[1].data["split_criterion"]
+                enumerate(possible_splits), key=lambda x: x[1].data["split_criterion"]
             ):
                 if i[1].data["split_permission"]:
                     min_density_node = i[1].identifier
@@ -299,7 +299,7 @@ class DePDDP:
 
         """
 
-        projection = None  # (ST2)
+        proj = None  # (ST2)
         splitpoint = None  # (ST2)
         split_criterion = None  # (ST2)
         flag = False  # (ST2)
@@ -307,13 +307,13 @@ class DePDDP:
         # if the number of samples
         if indices.shape[0] > self.min_sample_split:
             # execute pca on the data matrix
-            projection = util.execute_decomposition_method(
+            proj = util.execute_decomposition_method(
                 data_matrix=self.X[indices, :],
                 decomposition_method=self.decomposition_method,
                 two_dimentions=self.visualization_utility,
                 decomposition_args=self.decomposition_args,
             )
-            one_dimension = projection[:, 0]
+            one_dimension = proj[:, 0]
 
             # calculate the standard deviation of the data
             bandwidth = sm.nonparametric.bandwidths.select_bandwidth(
@@ -360,7 +360,7 @@ class DePDDP:
 
         return {
             "indices": indices,
-            "projection": projection,
+            "projection": proj,
             "splitpoint": splitpoint,
             "split_criterion": split_criterion,
             "split_permission": flag,
@@ -764,17 +764,17 @@ class IPDDP:
         # =========================
         if indices.shape[0] > self.min_sample_split:
             # execute pca on the data matrix
-            projection = util.execute_decomposition_method(
+            proj_vector = util.execute_decomposition_method(
                 data_matrix=self.X[indices, :],
                 decomposition_method=self.decomposition_method,
                 two_dimentions=self.visualization_utility,
                 decomposition_args=self.decomposition_args,
             )
-            one_dimension = projection[:, 0]
+            one_dimension = proj_vector[:, 0]
 
             sort_indices = np.argsort(one_dimension)
-            projection = projection[sort_indices, :]
-            one_dimension = projection[:, 0]
+            proj_vector = proj_vector[sort_indices, :]
+            one_dimension = proj_vector[:, 0]
             indices = indices[sort_indices]
 
             quantile_value = np.quantile(
@@ -801,20 +801,32 @@ class IPDDP:
                 flag = False
         # =========================
         else:
-            projection = None
+            proj_vector = None
             splitpoint = None
             split_criterion = None
             flag = False
 
         return {
             "indices": indices,
-            "projection": projection,
[…]
