+++ docs/HiPart.rst
@@ -1,11 +1,8 @@
-Documentation
-=============
+Package contents
+================
 
 .. toctree::
    :hidden:
-
-Package contents
----------------
 
 .. automodule:: HiPart
    :members:
