+++ paper/paper.md
@@ -1,5 +1,9 @@
 ---
 title: "HiPart: Hierarchical Divisive Clustering Toolbox"
+format:
+  pdf:
+    number-sections: true
+    colorlinks: true
 tags:
   - Python
   - Clustering
@@ -45,7 +49,7 @@ Although the discovery of a hierarchical format is crucial in many fields, such 
 
 Simultaneously, we provide implementations of a complete set of hierarchical divisive clustering algorithms with a similar basis. These are the dePDDP [@TASOULIS20103391], the iPDDP [@TASOULIS20103391], the kM-PDDP [@zeimpekis2008principal], and the bisecting k-Means (BKM) [@savaresi2001performance]. We also provide additional features not included in the original developments of the aforementioned methodologies that make them appropriate for the discovery of arbitrary shaped or non-linear separable clusters. In detail, we incorporate kernel Principal Component Analysis (kPCA) [@Scholkopf99kernelprincipal] and Independent Component Analysis (ICA) [@hyvarinen2000independent; @tharwat2020independent] for the iterative dimensionality reduction steps.
 
-As a result, the package provides a fully parameterized set of algorithms that can be applied in a diverse set of applications, for example, non-linear separable clusters, automated identification for the cluster number, and outlier control.
+As a result, the package provides a fully parameterized set of algorithms that can be applied in a diverse set of applications.
 
 
 # Software Description
@@ -59,7 +63,7 @@ The HiPart (Hierarchical Partitioning) package is divided into three major secti
 
 ## Method Implementation
 
-The package employs an object-oriented approach for the implementation of the algorithms, similarly to that of [@JMLR:v23:21-0862], while incorporating design similarities with the scikit-learn library [@pedregosa2011scikit]. Meaning, a class instance executes each of the algorithms, and the class instance's attributes are the algorithm's hyper-parameters and results.
+The package employs an object-oriented approach for the implementation of the algorithms, similarly to that of [@Philipp2022DoubleML], while incorporating design similarities with the scikit-learn library [@pedregosa2011scikit]. Meaning, a class instance executes each of the algorithms, and the class instance's attributes are the algorithm's hyper-parameters and results.
 
 For the execution of the algorithms, the user needs to call either the `predict` or the `fit_predict` method of each algorithm's execution class. The algorithm parameterization can be applied at the constructor of their respective class.
 
@@ -68,7 +72,7 @@ For the execution of the algorithms, the user needs to call either the `predict`
 
 Two static visualization methods are included. The first one is a 2-Dimensional representation of all the data splits generated by each algorithm during the hierarchical procedure. The goal is to provide an insight to the user regarding each node of the clustering tree and, subsequently, each step of the algorithm's execution.
 
-The second visualization method is a dendrogram that represents the splits of all the divisive algorithms. The dendrogram's figure creation is implemented by the *SciPy* package, and it is fully parameterized as stated in the library.
+The second visualization method is a dendrogram that represents the splits of all the divisive algorithms. The dendrogram's figure creation is implemented by the *SciPy* package [@2020SciPyNMeth], and it is fully parameterized as stated in the library.
 
 
 ## Interactive Visualization
@@ -100,20 +104,19 @@ Four widely used data sets from the field of bioinformatics are employed along w
 
 -   the Deng [@DengData],
 -   the Baron [@baron2016single],
--   the TGCA Pan-cancer[^1] (Cancer),
+-   the TGCA Pan-cancer [@cancerdata] (Cancer),
 -   the Chen [@chen2017single],
 -   the USPS [@291440],
 -   the BBC [@greene06icml],
 :::
-[^1]: https://www.doi.org/10.7303/syn300013
+
[…]
