+++ paper/paper.bib
@@ -1,3 +1,38 @@
+@article{2020SciPyNMeth,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+            Kern, Robert and Larson, Eric and Carey, C J and
+            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+            Harris, Charles R. and Archibald, Anne M. and
+            Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+            {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2},
+}
+
+@article{cancerdata,
+  title     = "Software-based supplementary materials in {MAPEH} for enhanced
+               learning resource management and delivery system",
+  author    = "R. Rala, Jessa",
+  journal   = "APJAET - Journal Asia Pacific Journal of Advanced Education and
+               Technology",
+  publisher = "International Organization of Educators and Researchers Inc.",
+  month     =  jan,
+  year      =  2023
+}
+
 @article{hyvarinen2000independent,
     title={Independent component analysis: algorithms and applications},
     author={Hyv{\"a}rinen, Aapo and Oja, Erkki},
@@ -39,7 +74,7 @@
     publisher={JMLR. org}
 }
 
-@article{JMLR:v23:21-0862,
+@article{Philipp2022DoubleML,
     author  = {Philipp Bach and Victor Chernozhukov and Malte S. Kurz and Martin Spindler},
     title   = {DoubleML - An Object-Oriented Implementation of Double Machine Learning in Python},
     journal = {Journal of Machine Learning Research},
@@ -330,15 +365,6 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
     pages = {357--362},
     doi = {10.1038/s41586-020-2649-2},
     publisher = {Springer Science and Business Media {LLC}},
-}
-
-@article{2020SciPy-NMeth,
[…]
