### 2023-02-01 #5024 @panagiotisanagnostou
In order to address the comments made by @jjerphan, I will respond by quoting his questions and providing my answers in the section below.

> > * **Contribution and authorship**: Has the submitting author (@panagiotisanagnostou) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> It is clear that @panagiotisanagnostou is the main major contributor to HiPart. @stevestavropoulos also has contributed notable to HiPart.
>
> To validate this item I would like to know the rationale for other authors' presence given that authors are indicated as having contributed equally.

The list of authors came after careful consideration of their contribution to the quality of the work done. Apparently the submitting author (@panagiotisanagnostou) is maintaining the GitHub repository but there are a lot of work done prior to uploading the source code. Prof. Sotiris Tasoulis was the original developer, prototyping the divisive clustering concept, so his insights into developing the basic design of the package were invaluable. Dr. Dimitris Tasoulis provided invaluable insight in the debugging process as a developer with industry experience, while also contributing in the preparation of the submitted manuscript. Finally, Prof. Vassilis Plagianakos contributed mainly to the visualization tools of the HiPart package with solutions that allow great expandability. We also highlight that all authors have published articles regarding divisive clustering, and thus contributed through their deep understanding and expertise of the domain.

> > * **Substantial scholarly effort**: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> I think HiPart is rather small as a package with perfectible aspects. To me, HiPart currently do not meet all the criteria for being eligible to be published in JOSS. Yet, I think it is on track and that this submission is appreciable and legitimate.
>
> Perfectible aspects of HiPart are:
>
> -   its documentation: for now, there is a small number of undocumented examples. Notebook styles examples for all methods might improve HiPart approachability and adoption.
>
> -  names containing typos and not PEP8-compliant: as an example, I think `bicecting_kmeans` must be renamed `BisectingKmeans`
>
> -  lack comparisons with other libraries in the Scientific Python Ecosystem, in particular proof of functionality for the targeted problems (scalable methods for high dimensional data clustering) with respect to alternatives.
>
> -  code quality: there's currently a lot of duplication on some files (see the ones graded as F). Due to the size of the library, some more effort can bring it to a A grade on Codacy.

Thank you for legitimizing our submission to the JOSS. Our goal for the following weeks is to add more methods and expand the documentation significantly. In the meantime, according to your suggestion, we have included much more well-presented examples in the documentation.

As far as your concerns on the PEP8 compliance and the quality of code, as you can see in the recent commits, we have addressed most issues.

The algorithm members of the HiPart package are binary space partition tree-based algorithms. Specifically, each split the algorithms conduct can be represented by a node in an unbalanced binary tree. The final clustering of the data in all the five algorithms are the samples contained as data in the leaves of the binary tree.

Moreover, as mentioned in the [CONTRIBUTING.md]( https://github.com/panagiotisanagnostou/HiPart/blob/main/CONTRIBUTING.md#your-first-code-contribution) the algorithms have the same three member methods in their implementation. Those method are, one to split the node into two children, one for selecting the nest node to split and one to calculate the data each node will contain. This fact has the unfortunate effect of creating duplication in the code, especially if  there is a need to keep the same basic structure in a source code. Nevertheless, we tried to reduce the duplication errors as much as we can.

After extensive research in the Scientific Python Ecosystem, to the best of our research efforts, divisive clustering algorithmic implementations are missing. This fact limits our options for comparisons; Still, we now decided to include further comparisons with more generic clustering toolboxes.

> > **Functionality**: Have the functional claims of the software been confirmed?
> > **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
> >
> > * **Reproducibility**: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
>
> I unfortunately can't reproduce results (benchmarks and figures) using script in [`paper/paper_scripts`](https://github.com/panagiotisanagnostou/HiPart/tree/paper/paper_scripts). Some import paths are invalid and data is missing.
>
> To validate this item, please do check that those scripts work (I think using relative paths might help prevent issues).
>
> Moreover, from my reading of the implementations (which are mainly backed by NumPy), I hardly think implementations can be scalable or efficient for the targeted problems.
>
> I can't say if HiPart is completely functional, is performant, and if claims are backed by reproducible results.

Thank you for highlighting this problem. We have now included all files needed for reproduction with the [paper_results_generation.py](https://github.com/panagiotisanagnostou/HiPart/blob/paper/paper_scripts/paper_results_generation.py) and the [generate_paper_fig.py](https://github.com/panagiotisanagnostou/HiPart/blob/paper/paper_scripts/generate_paper_fig.py) files. As far as the paper_scripts we uploaded the data in the repository in a compressed form.

We have now clarified the scalability of the included implementations for the target problems. We now clearly mention that the package focuses on Large Scale data that fit in memory and that the key aspect which makes it possible is the pass-by-reference nature of the Python programming language. In addition, we have also indicated that HiPart provide algorithmic options that differ significantly in complexity.

> > * **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Guidelines for 2) and 3) are clear to me but guidelines for 1) are missing. Adding a simple contributing guide as `CONTRIBUTING.md` would suffice to me.

Thank you for the suggestion we created a  [CONTRIBUTING.md]( https://github.com/panagiotisanagnostou/HiPart/blob/main/CONTRIBUTING.md#your-first-code-contribution) file. We believe that it contains all the necessary information for someone that wants to contribute to the repository to do so. We hope it is up to community standards.

> > * **Quality of writing**: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
> See [panagiotisanagnostou/HiPart#23](https://github.com/panagiotisanagnostou/HiPart/pull/23) for remarks for this items.

I have added a response in the pull request @jjerphan created.
