### 2023-02-11 #5024 @jjerphan
### Last review items' clarification

As of [panagiotisanagnostou/HiPart@efb7859](https://github.com/panagiotisanagnostou/HiPart/commit/efb78593984f301ddf1ad5a0573fe9ae806bc3f6), I have now checked:

 - Contribution and authorship
 - Community guidelines
 - Quality of writing

I have not checked the following 4 items of the 21 items mainly because I still either have remark or need more time to assess it:

 - Substantial scholarly effort
 - Functionality
 - Performance
 - Reproducibility

---

Responding to https://github.com/openjournals/joss-reviews/issues/5024#issuecomment-1412414533:

**Contribution and authorship**

Authorships is now clear. I guess part of this paragraph can be included in the project's `README.md`.

**Substantial scholarly effort**

> Thank you for legitimizing our submission to the JOSS. Our goal for the following weeks is to add more methods and expand the documentation significantly. In the meantime, according to your suggestion, we have included much more well-presented examples in the documentation.
>
> As far as your concerns on the PEP8 compliance and the quality of code, as you can see in the recent commits, we have addressed most issues.

Thank you, for mentioning this recent work.

> The algorithm members of the HiPart package are binary space partition tree-based algorithms. Specifically, each split the algorithms conduct can be represented by a node in an unbalanced binary tree. The final clustering of the data in all the five algorithms are the samples contained as data in the leaves of the binary tree.
>
> Moreover, as mentioned in the [CONTRIBUTING.md](https://github.com/panagiotisanagnostou/HiPart/blob/main/CONTRIBUTING.md?rgh-link-date=2023-02-01T17%3A10%3A22Z#your-first-code-contribution) the algorithms have the same three member methods in their implementation. Those method are, one to split the node into two children, one for selecting the nest node to split and one to calculate the data each node will contain. This fact has the unfortunate effect of creating duplication in the code, especially if there is a need to keep the same basic structure in a source code. Nevertheless, we tried to reduce the duplication errors as much as we can.

Thank you, I need more time to validate those aspects.

> After extensive research in the Scientific Python Ecosystem, to the best of our research efforts, divisive clustering algorithmic implementations are missing. This fact limits our options for comparisons; Still, we now decided to include further comparisons with more generic clustering toolboxes.

Thank you, I need more time to validate this.

**Functionality, Performance and Reproducibility**

I still can't run the scripts:

```bash
 → python paper_scripts/generate_paper_fig.py
Traceback (most recent call last):
  File "/home/jjerphan/dev/HiPart/paper_scripts/generate_paper_fig.py", line 29, in <module>
    import __utilities as util
ModuleNotFoundError: No module named '__utilities'
~/dev/HiPart HiPart ± paper
 → python paper_scripts/generate_paper_fig.py
~/dev/HiPart 🌍 HiPart 🐍 ± paper
 → python paper_scripts/paper_results_generation.py
Traceback (most recent call last):
  File "/home/jjerphan/dev/HiPart/paper_scripts/paper_results_generation.py", line 26, in <module>
    import __utilities as util
ModuleNotFoundError: No module named '__utilities'
```

You need to fix the imports.

**Community guidelines**

> Thank you for the suggestion we created a [CONTRIBUTING.md](https://github.com/panagiotisanagnostou/HiPart/blob/main/CONTRIBUTING.md?rgh-link-date=2023-02-01T17%3A10%3A22Z#your-first-code-contribution) file. We believe that it contains all the necessary information for someone that wants to contribute to the repository to do so. We hope it is up to community standards.

Thank you, this now qualifies to me.

**Quality of writing**

> I have added a response in the pull request @jjerphan created.

Thank you, this now qualifies to me.
