### 2023-02-13 f260802
Fixies on paper scripts
