### 2023-02-13 #5024 @panagiotisanagnostou
## Response to [comment](https://github.com/openjournals/joss-reviews/issues/5024#issuecomment-1426671491):

> **Functionality, Performance and Reproducibility**
>
> I still can't run the scripts:
>
> ```shell
>  → python paper_scripts/generate_paper_fig.py
> Traceback (most recent call last):
>   File "/home/jjerphan/dev/HiPart/paper_scripts/generate_paper_fig.py", line 29, in <module>
>     import __utilities as util
> ModuleNotFoundError: No module named '__utilities'
> ~/dev/HiPart HiPart ± paper
>  → python paper_scripts/generate_paper_fig.py
> ~/dev/HiPart HiPart ± paper
>  → python paper_scripts/paper_results_generation.py
> Traceback (most recent call last):
>   File "/home/jjerphan/dev/HiPart/paper_scripts/paper_results_generation.py", line 26, in <module>
>     import __utilities as util
> ModuleNotFoundError: No module named '__utilities'
> ```
>
> You need to fix the imports.

Thanks for bringing this up. It seems that, the problem you encountered with the scripts is that the working directory for the Python interpreter during execution should be the location of the script. We have updated the code, and this is no longer the case. The scripts are now executable from every directory in your system.  In addition, the two required packages are installed automatically upon execution.

The results of the execution can be found in the folder that contains the scripts under the names “dendrogram.pdf” and “paper_results.csv”.

Thank you so much for your time.
