### 2023-02-13 #5024 @jbytecode
@jjerphan - In page (https://joss.readthedocs.io/en/latest/editing.html#overview-of-editorial-process) it is stressed that

> We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

of course, any exceptions and excuses are welcome.
