### 2023-02-18 #5024 @jjerphan
> Thanks for bringing this up. It seems that, the problem you encountered with the scripts is that the working directory for the Python interpreter during execution should be the location of the script. We have updated the code, and this is no longer the case. The scripts are now executable from every directory in your system. In addition, the two required packages are installed automatically upon execution.

@panagiotisanagnostou: I am afraid, but I am still meeting the problem using the latest revision of the `main` branch (i.e. 0af766c2896070cce7fa54059a0401e308160072) and the latest revision of `paper` branch (i.e. 85e1426c9e14f6dc084d3b2ec75dc37dc71472da). Can you provide the commit of the resolution?
