### 2023-02-19 #5024 @panagiotisanagnostou
@jjerphan, The commit of the resolution can be located at the last commit of the Enhancements (i.e. f26080243e75e8aff0a6e88817a8f8bf180afacb). In this commit, you will see changes in three files. The "__utilities.py", the "generate_paper_fig.py" and the "paper_results_generation.py". The changes are (i) the change in the interpreter working directory, (ii) the try-except block for automatic installation of the "fuzzy-c-means" package, and the "h5py" package.

Please provide me with the error report in order to try and find a suitable solution. I really need help seeing what the problem might be.
