### 2023-02-19 #5024 @jjerphan
OK, the problem was that I ran python in [isolated mode](https://docs.python.org/3/using/cmdline.html#cmdoption-I). Disabling this mode allows importing from `__utilities.py`. Generally, I think it is better to use relative import _à la_:

```python
from .local_submodule import an_interfaces
```

as it allows supporting this mode.

What do you think?
