### 2023-02-19 a5c2192
Import fix for isolated mode and minor changes on the paper scripts
