+++ paper_scripts/paper_results_generation.py
@@ -24,7 +24,11 @@ Results generator.
 """
 
 import os
+import sys
+
 os.chdir(os.path.dirname(os.path.abspath(__file__)))
+if os.path.dirname(os.path.abspath(__file__)) not in sys.path:
+    sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 
 import __utilities as util
 import csv
