### 2023-02-19 #5024 @panagiotisanagnostou
The main problem in this mode is the following, as stated in [isolated mode](https://docs.python.org/3/using/cmdline.html#cmdoption-I):

> In isolated mode [sys.path](https://docs.python.org/3/library/sys.html#sys.path) contains neither the script's directory nor the user's site-packages directory. All PYTHON* environment variables are ignored, too.

This resulted in the sys.path variable to contain the following paths:
 *  '/usr/lib/python310.zip'
 *  '/usr/lib/python3.10'
 *  '/usr/lib/python3.10/lib-dynload'
 *  '/home/user/git/HiPart/env/lib/python3.10/site-packages'

Unfortunately, according to the above paths, there is no way to access the `__utilities.py` script in a relative way. The solution I used for this issue is to manually add the path to the script I execute to the `sys.path` variable, with the help of the `__file__` dunder variable, and this solves the problem of the imports.

I hope that this is a satisfactory solution.
