### 2023-02-23 #5024 @jjerphan
There's no problem if the script is executed without the `isolated` mode.
