### 2023-03-03 #5024 @jjerphan
To me, [the given elements for substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) are met:

  - [x] Age of software (is this a well-established software project) / length of commit history.
  - [x] Number of authors.
  - [x] Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
  - [x] Whether the software has already been cited in academic papers.
  - [x] Whether the software is sufficiently useful that it is likely to be cited by your peer group.

Hence I am validating the last "Substantial scholarly effort" item. I have not time to provide room of improvements but think previous indications have helped already.

Thank you all for your patience.
