### 2023-03-30 #5024 @panagiotisanagnostou
@jbytecode I would like to include some new features in the package that are not documented in the paper I submitted. Can I do it while the package is under review?
Thank you in advance for your time.
