+++ paper/paper.bib
@@ -65,7 +65,7 @@
 }
 
 @article{pedregosa2011scikit,
-    title={Scikit-learn: Machine learning in Python},
+    title={Scikit-learn: Machine learning in {P}ython},
     author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
     journal={the Journal of machine Learning research},
     volume={12},
@@ -76,7 +76,7 @@
 
 @article{Philipp2022DoubleML,
     author  = {Philipp Bach and Victor Chernozhukov and Malte S. Kurz and Martin Spindler},
-    title   = {DoubleML - An Object-Oriented Implementation of Double Machine Learning in Python},
+    title   = {DoubleML - An Object-Oriented Implementation of Double Machine Learning in {P}ython},
     journal = {Journal of Machine Learning Research},
     year    = {2022},
     volume  = {23},
@@ -217,7 +217,7 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
   publisher={Nature Publishing Group}
 }
 
-@misc{Dua2019 ,
+@misc{Dua2019,
     author = "Dua, Dheeru and Graff, Casey",
     year = "2017",
     title = "{UCI} Machine Learning Repository",
@@ -225,12 +225,12 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
     institution = "University of California, Irvine, School of Information and Computer Sciences" 
 }
 
-@inproceedings{Lang95
-    ,author = "Ken Lang"
-    ,title = "Newsweeder: Learning to filter netnews"
-    ,year = 1995
-    ,booktitle = "Proceedings of the Twelfth International Conference on Machine Learning"
-    ,pages = "331-339"
+@inproceedings{Lang95,
+    author = {Ken Lang},
+    title = {Newsweeder: Learning to filter netnews},
+    year = {1995},
+    booktitle = {Proceedings of the Twelfth International Conference on Machine Learning},
+    pages = {331-339}
 }
 
 
@@ -266,25 +266,25 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
 }
 
 @article{ankerst1999optics,
-  title={{OPTICS}: Ordering points to identify the clustering structure},
-  author={Ankerst, Mihael and Breunig, Markus M and Kriegel, Hans-Peter and Sander, J{\"o}rg},
-  journal={ACM Sigmod record},
-  volume={28},
-  number={2},
-  pages={49--60},
-  year={1999},
-  publisher={ACM New York, NY, USA}
+  title = {{OPTICS}: Ordering points to identify the clustering structure},
+  author = {Ankerst, Mihael and Breunig, Markus M and Kriegel, Hans-Peter and Sander, J{\"o}rg},
+  journal = {ACM Sigmod record},
+  volume = {28},
+  number = {2},
+  pages = {49--60},
+  year = {1999},
+  publisher = {ACM New York, NY, USA}
 }
 
 @article{ackermann2014analysis,
-  title={Analysis of agglomerative clustering},
-  author={Ackermann, Marcel R and Bl{\"o}mer, Johannes and Kuntze, Daniel and Sohler, Christian},
-  journal={Algorithmica},
-  volume={69},
-  number={1},
-  pages={184--215},
-  year={2014},
-  publisher={Springer}
+  title = {Analysis of agglomerative clustering},
[…]
