+++ paper/paper.bib
@@ -290,7 +290,7 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
 @INPROCEEDINGS{Scholkopf99kernelprincipal,
     author = {Bernhard Scholkopf and Alexander Smola and Klaus-Robert Müller},
     title = {Kernel principal component analysis},
-    booktitle = {ADVANCES IN KERNEL METHODS - SUPPORT VECTOR LEARNING},
+    booktitle = {Advances {I}n {K}ernel {M}ethods - {S}upport {V}ector {L}earning},
     year = {1999},
     pages = {327--352},
     publisher = {MIT Press}
