### 2023-04-09 #5024 @jbytecode
@panagiotisanagnostou - I've just a pull request that includes some minor changes in both the manuscript and the bibliography. Please review the PR. If you agree with them, please merge.

https://github.com/panagiotisanagnostou/HiPart/pull/28
