### 2023-04-10 #5024 @jbytecode
@panagiotisanagnostou - Could you please try adding missing DOI's as the editorial bot suggests? Addition to this, a single DOI seems to be invalid (you can replace the http header part and just include the DOI info).
