+++ paper/paper.bib
@@ -22,15 +22,14 @@
   doi     = {10.1038/s41592-019-0686-2},
 }
 
-@article{cancerdata,
-  title     = "Software-based supplementary materials in {MAPEH} for enhanced
-               learning resource management and delivery system",
-  author    = "R. Rala, Jessa",
-  journal   = "APJAET - Journal Asia Pacific Journal of Advanced Education and
-               Technology",
-  publisher = "International Organization of Educators and Researchers Inc.",
-  month     =  jan,
-  year      =  2023
+@misc{cancerdata,
+  author    = {Rala, Jessa R.},
+  title     = {{Software-Based Supplementary Materials in {MAPEH} for Enhanced Learning Resource Management and Delivery System}},
+  month        = feb,
+  year         = 2023,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.7602956},
+  url          = {https://doi.org/10.5281/zenodo.7602956}
 }
 
 @article{hyvarinen2000independent,
@@ -52,7 +51,8 @@
     number={12},
     pages={1363--1372},
     year={2014},
-    publisher={American Thoracic Society}
+    publisher={American Thoracic Society},
+    doi={10.1164/rccm.201406-1099OC}
 }
 
 @inproceedings{luo2003hierarchical,
@@ -61,7 +61,7 @@
     booktitle={Third IEEE Symposium on Bioinformatics and Bioengineering, 2003. Proceedings.},
     pages={328--335},
     year={2003},
-    organization={IEEE}
+    organization={IEEE},
 }
 
 @article{pedregosa2011scikit,
@@ -71,7 +71,7 @@
     volume={12},
     pages={2825--2830},
     year={2011},
-    publisher={JMLR. org}
+    publisher={JMLR.org}
 }
 
 @article{Philipp2022DoubleML,
@@ -82,7 +82,7 @@
     volume  = {23},
     number  = {53},
     pages   = {1-6},
-url     = {http://jmlr.org/papers/v23/21-0862.html}
+    url = {http://jmlr.org/papers/v23/21-0862.html}
 }
 
 @article{xu2015comprehensive,
@@ -104,7 +104,8 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
     number={13},
     pages={3227--3241},
     year={2017},
-    publisher={Elsevier}
+    publisher={Elsevier},
+    doi={10.1016/j.celrep.2017.03.004}
 }
 
 @article{baron2016single,
@@ -115,7 +116,8 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
     number={4},
     pages={346--360},
     year={2016},
-    publisher={Elsevier}
+    publisher={Elsevier},
+    doi={10.1016/j.cels.2016.08.011}
 }
 
 @article{DengData,
@@ -145,7 +147,8 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
 	Pages = {377--384},
 	Publisher = {ACM Press},
 	Title = {Practical Solutions to the Problem of Diagonal Dominance in Kernel Document Clustering},
-	Year = {2006}
+	Year = {2006},
+    doi = {10.1145/1143844.1143892}
 }
 
 @article{hutzenthaler2020overcoming,
@@ -159,11 +162,16 @@ url     = {http://jmlr.org/papers/v23/21-0862.html}
   publisher={The Royal Society Publishing}
 }
 
-@misc{hofmeyr2019ppci,
-  title={{PPCI}: an {R} package for cluster identification using projection pursuit},
[…]
