+++ paper/paper.bib
@@ -162,7 +162,7 @@
   publisher={The Royal Society Publishing}
 }
 
-@article{RJ-2019-046,
+@article{hofmeyr2019ppci,
   author = {David P. Hofmeyr and Nicos G. Pavlidis},
   title = {{PPCI: an R Package for Cluster Identification using Projection Pursuit}},
   year = {2019},
