### 2023-04-11 #5024 @panagiotisanagnostou
@jbytecode I am happy to tell you that the I have done the final checks required by the post review process.

- The final release, after the end of the review process, for the HiPart package is the version v0.3.1

- The Zenodo archive DOI for the HiPart package v0.3.1 is: http://doi.org/10.5281/zenodo.7814113

- I committed the changes required for the missing/invalid DOIs.

- I did the final double checks from the list "Additional Author Tasks After Review is Complete"

I hope everything is in order now.
