### 2023-04-18 #5024 @jjerphan
Thank you for this message and for managing the edition of the JOSS, @arfon.

I am relatively glad of this first review. I will try to be a bit more reactive for the subsequent ones, if I have time to review any other submissions.
